methods/SAStainDiff/requirements.txt
albumentations==1.3.1
blobfile==2.0.2
matplotlib==3.7.2
mpi4py==3.1.4
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
scipy==1.10.1
setuptools==60.2.0
tensorflow==2.13.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.2
